# openapi/nova-browser-api.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "ClassificationValidationRequest": {
        "properties": {
          "positive_label": {
            "default": 1,
            "title": "Positive Label",
            "type": "integer"
          },
          "y_pred": {
            "items": {
              "type": "integer"
            },
            "title": "Y Pred",
            "type": "array"
          },
          "y_score": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y Score"
          },
          "y_true": {
            "items": {
              "type": "integer"
            },
            "title": "Y True",
            "type": "array"
          }
        },
        "required": [
          "y_true",
          "y_pred"
        ],
        "title": "ClassificationValidationRequest",
        "type": "object"
      },
      "ClassificationValidationResponse": {
        "properties": {
          "auc_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auc Label"
          },
          "balanced_accuracy": {
            "title": "Balanced Accuracy",
            "type": "number"
          },
          "cohens_kappa": {
            "title": "Cohens Kappa",
            "type": "number"
          },
          "commission_error": {
            "title": "Commission Error",
            "type": "number"
          },
          "confusion": {
            "$ref": "#/components/schemas/ConfusionCountsSchema"
          },
          "error_rate": {
            "title": "Error Rate",
            "type": "number"
          },
          "f1_score": {
            "title": "F1 Score",
            "type": "number"
          },
          "kappa_label": {
            "title": "Kappa Label",
            "type": "string"
          },
          "mcc": {
            "title": "Mcc",
            "type": "number"
          },
          "omission_error": {
            "title": "Omission Error",
            "type": "number"
          },
          "overall_accuracy": {
            "title": "Overall Accuracy",
            "type": "number"
          },
          "precision": {
            "title": "Precision",
            "type": "number"
          },
          "prevalence": {
            "title": "Prevalence",
            "type": "number"
          },
          "recall": {
            "title": "Recall",
            "type": "number"
          },
          "roc_auc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roc Auc"
          },
          "specificity": {
            "title": "Specificity",
            "type": "number"
          }
        },
        "required": [
          "confusion",
          "overall_accuracy",
          "precision",
          "recall",
          "specificity",
          "f1_score",
          "omission_error",
          "commission_error",
          "cohens_kappa",
          "kappa_label",
          "balanced_accuracy",
          "mcc",
          "error_rate",
          "prevalence",
          "roc_auc",
          "auc_label"
        ],
        "title": "ClassificationValidationResponse",
        "type": "object"
      },
...
      "ConfusionCountsSchema": {
        "properties": {
          "fn": {
            "title": "Fn",
            "type": "integer"
          },
          "fp": {
            "title": "Fp",
            "type": "integer"
          },
          "n": {
            "title": "N",
            "type": "integer"
          },
          "tn": {
            "title": "Tn",
            "type": "integer"
          },
          "tp": {
            "title": "Tp",
            "type": "integer"
          }
        },
        "required": [
          "tp",
          "tn",
          "fp",
          "fn",
          "n"
        ],
        "title": "ConfusionCountsSchema",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "detail": {
            "title": "Detail"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "detail",
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "RegressionValidationRequest": {
        "properties": {
          "observed": {
            "items": {
              "type": "number"
            },
            "title": "Observed",
            "type": "array"
          },
          "predicted": {
            "items": {
              "type": "number"
            },
            "title": "Predicted",
            "type": "array"
          }
        },
        "required": [
          "observed",
          "predicted"
        ],
        "title": "RegressionValidationRequest",
        "type": "object"
      },
      "RegressionValidationResponse": {
        "properties": {
          "evs": {
            "title": "Evs",
            "type": "number"
          },
          "mae": {
            "title": "Mae",
            "type": "number"
          },
          "mape": {
            "title": "Mape",
            "type": "number"
          },
          "mbe": {
            "title": "Mbe",
            "type": "number"
          },
          "me": {
            "title": "Me",
            "type": "number"
          },
          "n": {
            "title": "N",
            "type": "integer"
          },
          "nse": {
            "title": "Nse",
            "type": "number"
          },
          "r_squared": {
            "title": "R Squared",
            "type": "number"
          },
          "rmse": {
            "title": "Rmse",
            "type": "number"
          },
          "rmse_quality_label": {
            "title": "Rmse Quality Label",
            "type": "string"
          },
          "rrmse": {
            "title": "Rrmse",
            "type": "number"
          },
          "se": {
            "title": "Se",
            "type": "number"
          },
          "see": {
            "title": "See",
            "type": "number"
          },
          "willmott_d": {
            "title": "Willmott D",
            "type": "number"
          }
        },
        "required": [
          "n",
          "me",
          "mbe",
          "mae",
          "rmse",
          "mape",
          "se",
          "nse",
          "r_squared",
          "see",
          "rrmse",
          "willmott_d",
          "evs",
          "rmse_quality_label"
        ],
        "title": "RegressionValidationResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "NOVA-GeoRisk Browser API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/validation/classification": {
      "post": {
        "operationId": "validate_classification_api_v1_validation_classification_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassificationValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificationValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Validate Classification",
        "tags": [
          "Validation"
        ]
      }
    },
    "/api/v1/validation/regression": {
      "post": {
        "operationId": "validate_regression_api_v1_validation_regression_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegressionValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegressionValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Validate Regression",
        "tags": [
          "Validation"
        ]
      }
    },
...
  }
}